Add experiments for revised manuscript
Add ADMM-PINNs for the optimal control of Burgers equation.
Add SSN-FEM and ADMM-FEM for the optimal control of bilinear parabolic
  equation.
Add SSN-FEM and PD-FEM for the optimal control of semilinear parabolic
  equation.
Remove FNO model parameters for the semilinear parabolic problem due to
  their large file size.
Format, restructure, and clean up the repository for better readability.